🌟 [Major]: Add `Path` parameter to replace `WorkingDirectory` (#3)
## Description

This pull request includes changes to the `README.md` and `action.yml`
files to update the artifact download process and correct the path
usage. This includes the renaming of the `WorkingDirectory` input to
`Path`, updating the `README.md` to reflect this change, and modifying
the script path in the `action.yml`.

Updates to artifact download process:

*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5L12-R17):
Renamed `WorkingDirectory` input to `Path` and updated the description
to reflect the new input name.
*
[`action.yml`](diffhunk://#diff-1243c5424efaaa19bd8e813c5e6f6da46316e63761421b3e5f5c8ced9a36e6b6L9-R11):
Renamed `WorkingDirectory` input to `Path` and updated the description
accordingly.
[[1]](diffhunk://#diff-1243c5424efaaa19bd8e813c5e6f6da46316e63761421b3e5f5c8ced9a36e6b6L9-R11)
[[2]](diffhunk://#diff-1243c5424efaaa19bd8e813c5e6f6da46316e63761421b3e5f5c8ced9a36e6b6R28-R30)

Path usage correction:

*
[`action.yml`](diffhunk://#diff-1243c5424efaaa19bd8e813c5e6f6da46316e63761421b3e5f5c8ced9a36e6b6L42-R51):
Modified the script path from backslashes to forward slashes for
compatibility.

## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [ ] 🪲 [Fix]
- [ ] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [ ] 🚀 [Feature]
- [x] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas